Make `Grib2::submessages()` return an iterator instead of a boxed slice
From each element of the original slice, only the index information of sections
in each submessage could be accessed, but from each element of the new iterator,
other information of each submessage can be accessed.  Therefore, this
`Grib2::submessages()` is now important and plays a fundamental role for users
who want to access any submessage information.