Allow messages to be read without error even if non-GRIB byte sequences exist in-between
The process of reading Section 0 has been rewritten from scratch so that
the next message is read even if an arbitrary byte sequence is inserted
before the message.

This change has changed the behavior of the reader, as the byte sequence
given to the reader is no longer expected to start with `GRIB`, so
errors caused by failure to read `GRIB` no longer occur.