Correct support for cases where nbit is 0 in the complex packing decoder
The logic for decoding complex packing when nbit is 0 was incorrect and
wrong values were output.  This is due to the fact that I added support
for "nbit=0" cases in 26dd5dd4b06a19808fa85a6f9b1ff565bce6725e based on
a wrong understanding and tests happened to pass with the data used for
confirmation.

This time I read the source code of wgrib2 and revised my understanding
of the specification. Thanks wgrib2!